+++ db/migrations/20251126083933_run_pid.sql
@@ -1,0 +1,7 @@
+-- migrate:up
+alter table runs
+add column pid int default null;
+
+-- migrate:down
+alter table runs
+drop column pid;