+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Displaying of emojis using config value `display.emoji`. Defaults to `true`.
 - Colour run status output using config value `display.color.status.[succeeded|failed|running|skipped|terminated]`. Defaults to `false`.
 - Handling SIGTERM in `exec` by setting run state to `Terminated`.
+- `run kill` to kill runs with status `Running`.
 
 ## [0.2.0] - 2025-11-23